Decompose Gradle checks into convention plugins
