+++ build-logic/convention/src/main/kotlin/vibits.checks.coverage.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+  id("org.jetbrains.kotlinx.kover")
+}
+
+dependencies {
+  subprojects.forEach { subproject ->
+    subproject.plugins.withId("org.jetbrains.kotlinx.kover") {
+      kover(subproject)
+    }
+  }
+}
